+++ beets/autotag/mb.py
@@ -48,15 +48,10 @@ RELEASE_INCLUDES = ['artists', 'media', 'recordings', 'release-groups',
                     'labels', 'artist-credits']
 TRACK_INCLUDES = ['artists']
 
-# python-musicbrainz-ngs search functions: tolerate different API versions.
-if hasattr(musicbrainzngs, 'release_search'):
-    # Old API names.
-    _mb_release_search = musicbrainzngs.release_search
-    _mb_recording_search = musicbrainzngs.recording_search
-else:
-    # New API names.
-    _mb_release_search = musicbrainzngs.search_releases
-    _mb_recording_search = musicbrainzngs.search_recordings
+# Only versions >= 0.3 of python-musicbrainz-ngs support artist aliases.
+if musicbrainzngs.musicbrainz._version >= '0.3':
+    RELEASE_INCLUDES.append('aliases')
+    TRACK_INCLUDES.append('aliases')
 
 def configure():
     """Set up the python-musicbrainz-ngs module according to settings
@@ -67,6 +62,34 @@ def configure():
         config['musicbrainz']['ratelimit_interval'].as_number(),
         config['musicbrainz']['ratelimit'].get(int),
     )
+
+def _preferred_alias(aliases):
+    """Given an list of alias structures for an artist credit, select
+    and return the user's preferred alias alias or None if no matching
+    alias is found.
+    """
+    if not aliases:
+        return
+
+    # Only consider aliases that have locales set.
+    aliases = [a for a in aliases if 'locale' in a]
+
+    # Search configured locales in order.
+    for locale in config['import']['languages'].as_str_seq():
+        # Find matching aliases for this locale.
+        matches = [a for a in aliases if a['locale'] == locale]
+        # Skip to the next locale if we have no matches
+        if not matches:
+            continue
+
+        # Find the aliases that have the primary flag set.
+        primaries = [a for a in matches if 'primary' in a]
+        # Take the primary if we have it, otherwise take the first
+        # match with the correct locale.
+        if primaries:
+            return primaries[0]
+        else:
+            return matches[0]
 
 def _flatten_artist_credit(credit):
     """Given a list representing an ``artist-credit`` block, flatten the
@@ -84,12 +107,19 @@ def _flatten_artist_credit(credit):
             artist_sort_parts.append(el)
 
         else:
+            alias = _preferred_alias(el['artist'].get('alias-list', ()))
+
             # An artist.
-            cur_artist_name = el['artist']['name']
+            if alias:
+                cur_artist_name = alias['alias']
+            else:
+                cur_artist_name = el['artist']['name']
             artist_parts.append(cur_artist_name)
 
             # Artist sort name.
-            if 'sort-name' in el['artist']:
+            if alias:
+                artist_sort_parts.append(alias['sort-name'])
+            elif 'sort-name' in el['artist']:
                 artist_sort_parts.append(el['artist']['sort-name'])
             else:
                 artist_sort_parts.append(cur_artist_name)
@@ -174,7 +204,7 @@ def album_info(release):
                             int(medium['position']),
                             int(track['position']))
             if track.get('title'):
-                # Track title may be distinct from underling recording
+                # Track title may be distinct from underlying recording
                 # title.
                 ti.title = track['title']
             ti.disctitle = disctitle
@@ -264,7 +294,7 @@ def match_album(artist, album, tracks=None, limit=SEARCH_LIMIT):
         return
 
     try:
-        res = _mb_release_search(limit=limit, **criteria)
+        res = musicbrainzngs.search_releases(limit=limit, **criteria)
     except musicbrainzngs.MusicBrainzError as exc:
         raise MusicBrainzAPIError(exc, 'release search', criteria,
                                   traceback.format_exc())
@@ -288,7 +318,7 @@ def match_track(artist, title, limit=SEARCH_LIMIT):
         return
 
     try:
-        res = _mb_recording_search(limit=limit, **criteria)
+        res = musicbrainzngs.search_recordings(limit=limit, **criteria)
     except musicbrainzngs.MusicBrainzError as exc:
         raise MusicBrainzAPIError(exc, 'recording search', criteria,
                                   traceback.format_exc())