+++ beetsplug/echonest_tempo.py
@@ -59,6 +59,8 @@ def get_tempo(artist, title):
     """Get the tempo for a song."""
     # We must have sufficient metadata for the lookup. Otherwise the API
     # will just complain.
+    artist = artist.replace(u'\n', u' ').strip()
+    title = title.replace(u'\n', u' ').strip()
     if not artist or not title:
         return None
 
@@ -75,7 +77,8 @@ def get_tempo(artist, title):
                 # Wait and try again.
                 time.sleep(RETRY_INTERVAL)
             else:
-                raise
+                log.warn(u'echonest_tempo: {0}'.format(e.args[0][0]))
+                return None
         except pyechonest.util.EchoNestIOError as e:
             log.debug(u'echonest_tempo: IO error: {0}'.format(e))
             time.sleep(RETRY_INTERVAL)