+++ beetsplug/smartplaylist.py
@@ -42,7 +42,8 @@ def update_playlists(lib):
         # As we allow tags in the m3u names, we'll need to iterate through
         # the items and generate the correct m3u file names.
         for item in items:
-            m3u_name = item.evaluate_template(Template(basename), lib=lib)
+            m3u_name = item.evaluate_template(Template(basename), lib=lib,
+                sanitize=True)
             if not (m3u_name in m3us):
                 m3us[m3u_name] = []
             if relative_to: