+++ test/_common.py
@@ -17,8 +17,8 @@ import time
 import sys
 import os
 import logging
-import contextlib
-import copy
+import tempfile
+import shutil
 
 # Use unittest2 on Python < 2.7.
 try:
@@ -80,20 +80,40 @@ def import_session(lib=None, logfile=None, paths=[], query=[], cli=False):
     cls = commands.TerminalImportSession if cli else importer.ImportSession
     return cls(lib, logfile, paths, query)
 
-# Temporary config modifications.
+# A test harness for all beets tests.
+# Provides temporary, isolated configuration.
 class TestCase(unittest.TestCase):
     """A unittest.TestCase subclass that saves and restores beets'
     global configuration. This allows tests to make temporary
     modifications that will then be automatically removed when the test
-    completes. Also provides some additional assertion methods.
+    completes. Also provides some additional assertion methods, a
+    temporary directory, and a DummyIO.
     """
     def setUp(self):
         # A "clean" source list including only the defaults.
         beets.config.sources = []
         beets.config.read(user=False, defaults=True)
 
+        # Direct paths to a temporary directory. Tests can also use this
+        # temporary directory.
+        self.temp_dir = tempfile.mkdtemp()
+        beets.config['statefile'] = os.path.join(self.temp_dir, 'state.pickle')
+        beets.config['library'] = os.path.join(self.temp_dir, 'library.db')
+        beets.config['directory'] = os.path.join(self.temp_dir, 'libdir')
+
+        # Set $HOME, which is used by confit's `config_dir()` to create
+        # directories.
+        self._old_home = os.environ.get('HOME')
+        os.environ['HOME'] = self.temp_dir
+
+        # Initialize, but don't install, a DummyIO.
+        self.io = DummyIO()
+
     def tearDown(self):
-        pass
+        if os.path.isdir(self.temp_dir):
+            shutil.rmtree(self.temp_dir)
+        os.environ['HOME'] = self._old_home
+        self.io.restore()
 
     def assertExists(self, path):
         self.assertTrue(os.path.exists(path),