+++ test/test_art.py
@@ -50,12 +50,11 @@ class FetchImageTest(unittest.TestCase):
         artpath = fetchart._fetch_image('http://example.com')
         self.assertNotEqual(artpath, None)
 
-class FSArtTest(unittest.TestCase):
+class FSArtTest(_common.TestCase):
     def setUp(self):
-        self.dpath = os.path.join(_common.RSRC, 'arttest')
+        super(FSArtTest, self).setUp()
+        self.dpath = os.path.join(self.temp_dir, 'arttest')
         os.mkdir(self.dpath)
-    def tearDown(self):
-        shutil.rmtree(self.dpath)
 
     def test_finds_jpg_in_directory(self):
         _common.touch(os.path.join(self.dpath, 'a.jpg'))
@@ -73,16 +72,22 @@ class FSArtTest(unittest.TestCase):
         fn = fetchart.art_in_path(self.dpath)
         self.assertEqual(fn, None)
 
-class CombinedTest(unittest.TestCase):
+class CombinedTest(_common.TestCase):
     def setUp(self):
-        self.dpath = os.path.join(_common.RSRC, 'arttest')
+        super(CombinedTest, self).setUp()
+
+        self.dpath = os.path.join(self.temp_dir, 'arttest')
         os.mkdir(self.dpath)
         self.old_urlopen = fetchart.urllib.urlopen
         fetchart.urllib.urlopen = self._urlopen
         self.page_text = ""
         self.urlopen_called = False
+
+        # Set up configuration.
+        fetchart.FetchArtPlugin()
+
     def tearDown(self):
-        shutil.rmtree(self.dpath)
+        super(CombinedTest, self).tearDown()
         fetchart.urllib.urlopen = self.old_urlopen
 
     def _urlopen(self, url):
@@ -184,7 +189,7 @@ class ArtImporterTest(_common.TestCase):
         super(ArtImporterTest, self).setUp()
 
         # Mock the album art fetcher to always return our test file.
-        self.art_file = os.path.join(_common.RSRC, 'tmpcover.jpg')
+        self.art_file = os.path.join(self.temp_dir, 'tmpcover.jpg')
         _common.touch(self.art_file)
         self.old_afa = fetchart.art_for_album
         self.afa_response = self.art_file
@@ -193,8 +198,8 @@ class ArtImporterTest(_common.TestCase):
         fetchart.art_for_album = art_for_album
 
         # Test library.
-        self.libpath = os.path.join(_common.RSRC, 'tmplib.blb')
-        self.libdir = os.path.join(_common.RSRC, 'tmplib')
+        self.libpath = os.path.join(self.temp_dir, 'tmplib.blb')
+        self.libdir = os.path.join(self.temp_dir, 'tmplib')
         os.mkdir(self.libdir)
         os.mkdir(os.path.join(self.libdir, 'album'))
         itempath = os.path.join(self.libdir, 'album', 'test.mp3')
@@ -224,14 +229,7 @@ class ArtImporterTest(_common.TestCase):
 
     def tearDown(self):
         super(ArtImporterTest, self).tearDown()
-
         fetchart.art_for_album = self.old_afa
-        if os.path.exists(self.art_file):
-            os.remove(self.art_file)
-        if os.path.exists(self.libpath):
-            os.remove(self.libpath)
-        if os.path.exists(self.libdir):
-            shutil.rmtree(self.libdir)
 
     def _fetch_art(self, should_exist):
         """Execute the fetch_art coroutine for the task and return the