+++ test/test_mediafile.py
@@ -189,6 +189,7 @@ class ZeroLengthMediaFile(beets.mediafile.MediaFile):
         return 0.0
 class MissingAudioDataTest(unittest.TestCase):
     def setUp(self):
+        super(MissingAudioDataTest, self).setUp()
         path = os.path.join(_common.RSRC, 'full.mp3')
         self.mf = ZeroLengthMediaFile(path)
 
@@ -198,12 +199,48 @@ class MissingAudioDataTest(unittest.TestCase):
 
 class TypeTest(unittest.TestCase):
     def setUp(self):
+        super(TypeTest, self).setUp()
         path = os.path.join(_common.RSRC, 'full.mp3')
         self.mf = beets.mediafile.MediaFile(path)
 
     def test_year_integer_in_string(self):
         self.mf.year = '2009'
         self.assertEqual(self.mf.year, 2009)
+
+    def test_set_replaygain_gain_to_none(self):
+        self.mf.rg_track_gain = None
+        self.assertEqual(self.mf.rg_track_gain, 0.0)
+
+    def test_set_replaygain_peak_to_none(self):
+        self.mf.rg_track_peak = None
+        self.assertEqual(self.mf.rg_track_peak, 0.0)
+
+    def test_set_year_to_none(self):
+        self.mf.year = None
+        self.assertEqual(self.mf.year, 0)
+
+    def test_set_track_to_none(self):
+        self.mf.track = None
+        self.assertEqual(self.mf.track, 0)
+
+class SoundCheckTest(unittest.TestCase):
+    def test_round_trip(self):
+        data = beets.mediafile._sc_encode(1.0, 1.0)
+        gain, peak = beets.mediafile._sc_decode(data)
+        self.assertEqual(gain, 1.0)
+        self.assertEqual(peak, 1.0)
+
+    def test_decode_zero(self):
+        data = u' 80000000 80000000 00000000 00000000 00000000 00000000 ' \
+               u'00000000 00000000 00000000 00000000'
+        gain, peak = beets.mediafile._sc_decode(data)
+        self.assertEqual(gain, 0.0)
+        self.assertEqual(peak, 0.0)
+
+    def test_malformatted(self):
+        gain, peak = beets.mediafile._sc_decode(u'foo')
+        self.assertEqual(gain, 0.0)
+        self.assertEqual(peak, 0.0)
 
 def suite():
     return unittest.TestLoader().loadTestsFromName(__name__)